+++ modules/generator/overrides.go
@@ -1,8 +1,11 @@
 package generator
 
 import (
+	"time"
+
 	"github.com/grafana/tempo/modules/generator/registry"
 	"github.com/grafana/tempo/modules/overrides"
+	"github.com/grafana/tempo/pkg/sharedconfig"
 	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
@@ -12,10 +15,19 @@ type metricsGeneratorOverrides interface {
 	MetricsGeneratorProcessors(userID string) map[string]struct{}
 	MetricsGeneratorProcessorServiceGraphsHistogramBuckets(userID string) []float64
 	MetricsGeneratorProcessorServiceGraphsDimensions(userID string) []string
+	MetricsGeneratorProcessorServiceGraphsPeerAttributes(userID string) []string
 	MetricsGeneratorProcessorSpanMetricsHistogramBuckets(userID string) []float64
 	MetricsGeneratorProcessorSpanMetricsDimensions(userID string) []string
 	MetricsGeneratorProcessorSpanMetricsIntrinsicDimensions(userID string) map[string]bool
 	MetricsGeneratorProcessorSpanMetricsFilterPolicies(userID string) []filterconfig.FilterPolicy
+	MetricsGeneratorProcessorLocalBlocksMaxLiveTraces(userID string) uint64
+	MetricsGeneratorProcessorLocalBlocksMaxBlockDuration(userID string) time.Duration
+	MetricsGeneratorProcessorLocalBlocksMaxBlockBytes(userID string) uint64
+	MetricsGeneratorProcessorLocalBlocksTraceIdlePeriod(userID string) time.Duration
+	MetricsGeneratorProcessorLocalBlocksFlushCheckPeriod(userID string) time.Duration
+	MetricsGeneratorProcessorLocalBlocksCompleteBlockTimeout(userID string) time.Duration
+	MetricsGeneratorProcessorSpanMetricsDimensionMappings(userID string) []sharedconfig.DimensionMappings
+	MetricsGeneratorProcessorSpanMetricsEnableTargetInfo(userID string) bool
 }
 
 var _ metricsGeneratorOverrides = (*overrides.Overrides)(nil)