+++ modules/generator/processor/servicegraphs/servicegraphs.go
@@ -5,6 +5,7 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/go-kit/log"
@@ -13,6 +14,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/prometheus/util/strutil"
+	"go.opentelemetry.io/otel/attribute"
+	semconv "go.opentelemetry.io/otel/semconv/v1.18.0"
 
 	gen "github.com/grafana/tempo/modules/generator/processor"
 	"github.com/grafana/tempo/modules/generator/processor/servicegraphs/store"
@@ -49,6 +52,10 @@ const (
 	metricRequestClientSeconds = "traces_service_graph_request_client_seconds"
 )
 
+var defaultPeerAttributes = []attribute.Key{
+	semconv.PeerServiceKey, semconv.NetPeerNameKey, semconv.NetSockPeerNameKey, semconv.RPCServiceKey, semconv.NetSockPeerAddrKey, semconv.HTTPURLKey, semconv.HTTPTargetKey,
+}
+
 type tooManySpansError struct {
 	droppedSpans int
 }
@@ -61,6 +68,7 @@ type Processor struct {
 	Cfg Config
 
 	registry registry.Registry
+	labels   []string
 	store    store.Store
 
 	closeCh chan struct{}
@@ -85,13 +93,13 @@ func New(cfg Config, tenant string, registry registry.Registry, logger log.Logge
 	p := &Processor{
 		Cfg:      cfg,
 		registry: registry,
+		labels:   labels,
+		closeCh:  make(chan struct{}, 1),
 
-		closeCh: make(chan struct{}, 1),
-
-		serviceGraphRequestTotal:                  registry.NewCounter(metricRequestTotal, labels),
-		serviceGraphRequestFailedTotal:            registry.NewCounter(metricRequestFailedTotal, labels),
-		serviceGraphRequestServerSecondsHistogram: registry.NewHistogram(metricRequestServerSeconds, labels, cfg.HistogramBuckets),
-		serviceGraphRequestClientSecondsHistogram: registry.NewHistogram(metricRequestClientSeconds, labels, cfg.HistogramBuckets),
+		serviceGraphRequestTotal:                  registry.NewCounter(metricRequestTotal),
+		serviceGraphRequestFailedTotal:            registry.NewCounter(metricRequestFailedTotal),
+		serviceGraphRequestServerSecondsHistogram: registry.NewHistogram(metricRequestServerSeconds, cfg.HistogramBuckets),
+		serviceGraphRequestClientSecondsHistogram: registry.NewHistogram(metricRequestClientSeconds, cfg.HistogramBuckets),
 
 		metricDroppedSpans: metricDroppedSpans.WithLabelValues(tenant),
 		metricTotalEdges:   metricTotalEdges.WithLabelValues(tenant),
@@ -168,6 +176,7 @@ func (p *Processor) consume(resourceSpans []*v1_trace.ResourceSpans) (err error)
 						e.Failed = e.Failed || p.spanFailed(span)
 						p.upsertDimensions(e.Dimensions, rs.Resource.Attributes, span.Attributes)
 						e.SpanMultiplier = spanMultiplier
+						p.upsertPeerNode(e, span.Attributes)
 
 						// A database request will only have one span, we don't wait for the server
 						// span but just copy details from the client span
@@ -192,6 +201,7 @@ func (p *Processor) consume(resourceSpans []*v1_trace.ResourceSpans) (err error)
 						e.Failed = e.Failed || p.spanFailed(span)
 						p.upsertDimensions(e.Dimensions, rs.Resource.Attributes, span.Attributes)
 						e.SpanMultiplier = spanMultiplier
+						p.upsertPeerNode(e, span.Attributes)
 					})
 				default:
 					// this span is not part of an edge
@@ -233,6 +243,15 @@ func (p *Processor) upsertDimensions(m map[string]string, resourceAttr []*v1_com
 	}
 }
 
+func (p *Processor) upsertPeerNode(e *store.Edge, spanAttr []*v1_common.KeyValue) {
+	for _, peerKey := range p.Cfg.PeerAttributes {
+		if v, ok := processor_util.FindAttributeValue(peerKey, spanAttr); ok {
+			e.PeerNode = v
+			return
+		}
+	}
+}
+
 func (p *Processor) Shutdown(_ context.Context) {
 	close(p.closeCh)
 }
@@ -245,7 +264,7 @@ func (p *Processor) onComplete(e *store.Edge) {
 		labelValues = append(labelValues, e.Dimensions[dimension])
 	}
 
-	registryLabelValues := p.registry.NewLabelValues(labelValues)
+	registryLabelValues := p.registry.NewLabelValueCombo(p.labels, labelValues)
 
 	p.serviceGraphRequestTotal.Inc(registryLabelValues, 1*e.SpanMultiplier)
 	if e.Failed {
@@ -258,6 +277,25 @@ func (p *Processor) onComplete(e *store.Edge) {
 
 func (p *Processor) onExpire(e *store.Edge) {
 	p.metricExpiredEdges.Inc()
+
+	// If an edge is expired, we check if there are signs that the missing span is belongs to a "virtual node".
+	// These are nodes that are outside the user's reach (eg. an external service for payment processing),
+	// or that are not instrumented (eg. a frontend application).
+	e.ConnectionType = store.VirtualNode
+	if len(e.ClientService) == 0 {
+		// If the client service is not set, it means that the span could have been initiated by an external system,
+		// like a frontend application or an engineer via `curl`.
+		// We check if the span we have is the root span, and if so, we set the client service to "user".
+		if _, parentSpan := parseKey(e.Key()); len(parentSpan) == 0 {
+			e.ClientService = "user"
+			p.onComplete(e)
+		}
+	} else if len(e.ServerService) == 0 && len(e.PeerNode) > 0 {
+		// If client span does not have its matching server span, but has a peer attribute present,
+		// we make the assumption that a call was made to an external service, for which Tempo won't receive spans.
+		e.ServerService = e.PeerNode
+		p.onComplete(e)
+	}
 }
 
 func (p *Processor) spanFailed(span *v1_trace.Span) bool {
@@ -270,4 +308,9 @@ func spanDurationSec(span *v1_trace.Span) float64 {
 
 func buildKey(k1, k2 string) string {
 	return fmt.Sprintf("%s-%s", k1, k2)
+}
+
+func parseKey(key string) (string, string) {
+	parts := strings.Split(key, "-")
+	return parts[0], parts[1]
 }