+++ tempodb/encoding/common/interfaces.go
@@ -24,6 +24,7 @@ type Searcher interface {
 	Search(ctx context.Context, req *tempopb.SearchRequest, opts SearchOptions) (*tempopb.SearchResponse, error)
 	SearchTags(ctx context.Context, scope traceql.AttributeScope, cb TagCallback, opts SearchOptions) error
 	SearchTagValues(ctx context.Context, tag string, cb TagCallback, opts SearchOptions) error
+	SearchTagValuesV2(ctx context.Context, tag traceql.Attribute, cb TagCallbackV2, opts SearchOptions) error
 
 	Fetch(context.Context, traceql.FetchSpansRequest, SearchOptions) (traceql.FetchSpansResponse, error)
 }
@@ -94,6 +95,8 @@ type WALBlock interface {
 
 	// Append the given trace to the block. Must be safe for concurrent use with read operations.
 	Append(id ID, b []byte, start, end uint32) error
+
+	AppendTrace(id ID, tr *tempopb.Trace, start, end uint32) error
 
 	// Flush any unbuffered data to disk.  Must be safe for concurrent use with read operations.
 	Flush() error